Refactor buddy list to separate files
